Switch main menu coordinates to float

+++ src/core/Core.h
@@ -104,8 +104,8 @@ inline float minSizeRatio() { return std::min(g_sizeRatio.x, g_sizeRatio.y); }
 #define RATIO_X(a)    (((float)a)*g_sizeRatio.x)
 #define RATIO_Y(a)    (((float)a)*g_sizeRatio.y)
 
-inline Vec2i RATIO_2(const Vec2i & in) {
-	return Vec2i(RATIO_X(in.x), RATIO_Y(in.y));
+inline Vec2f RATIO_2(const Vec2f & in) {
+	return Vec2f(RATIO_X(in.x), RATIO_Y(in.y));
 }
