+++ src/gui/widget/CycleTextWidget.h
@@ -41,7 +41,7 @@ public:
 	
 	void AddText(TextWidget * text);
 	
-	void Move(const Vec2i & offset);
+	void Move(const Vec2f & offset);
 	bool OnMouseClick();
 	Widget * OnShortCut() { return NULL; }
 	void Update(int time);
